Split up weather into presentational and container components and updated background to be tinted
